+++ Engine/source/BasicComponents/Sprite.cpp
@@ -63,21 +63,21 @@ void Sprite::UpdateProjection()
 void Sprite::AddQuad(glm::vec4 color_)
 {
 	Engine::Instance().GetVKRenderManager()->LoadQuad(color_, 0.f, 0.f);
-	materialId = Engine::Instance().GetVKRenderManager()->GetVertexVector()->size() - 1;
+	materialId = static_cast<int>(Engine::Instance().GetVKRenderManager()->GetVertexVector()->size() - 1);
 	AddSpriteToManager();
 }
 
 void Sprite::AddQuadLine(glm::vec4 color_)
 {
 	Engine::Instance().GetVKRenderManager()->LoadLineQuad(color_);
-	materialId = Engine::Instance().GetVKRenderManager()->GetVertexVector()->size() - 1;
+	materialId = static_cast<int>(Engine::Instance().GetVKRenderManager()->GetVertexVector()->size() - 1);
 	AddSpriteToManager();
 }
 
 void Sprite::AddMeshWithTexture(std::string name_, glm::vec4 color_)
 {
 	Engine::Instance().GetVKRenderManager()->LoadQuad(color_, 1.f, 0.f);
-	materialId = Engine::Instance().GetVKRenderManager()->GetVertexVector()->size() - 1;
+	materialId = static_cast<int>(Engine::Instance().GetVKRenderManager()->GetVertexVector()->size() - 1);
 	ChangeTexture(name_);
 	textureSize = Engine::Instance().GetVKRenderManager()->GetTexture(name_)->GetSize();
 	AddSpriteToManager();
@@ -86,7 +86,7 @@ void Sprite::AddMeshWithTexture(std::string name_, glm::vec4 color_)
 void Sprite::AddMeshWithTexel(std::string name_, glm::vec4 color_)
 {
 	Engine::Instance().GetVKRenderManager()->LoadQuad(color_, 1.f, 1.f);
-	materialId = Engine::Instance().GetVKRenderManager()->GetVertexVector()->size() - 1;
+	materialId = static_cast<int>(Engine::Instance().GetVKRenderManager()->GetVertexVector()->size() - 1);
 	ChangeTexture(name_);
 	textureSize = Engine::Instance().GetVKRenderManager()->GetTexture(name_)->GetSize();
 	AddSpriteToManager();
